dev-4825 added more tests
